Carrito

+++ src/store/cartReducer.js
@@ -1,17 +1,29 @@
 import { createReducer, createAsyncThunk, createAction } from "@reduxjs/toolkit";
 import axios from "axios";
+import { addOneToQuantity, removeOneFromQuantity } from "../utils";
 
 export const getItems = createAsyncThunk(
   "GET_ITEMS",
   async (data, thunkAPI) => {
     try {
-      const { token, id } = data;
+      const { isLogIn } = thunkAPI.getState();
 
-      const req = await axios.get(`/api/cart/items/${id}`, {
-        headers: { authorization: token },
-      });
+      if (isLogIn) {
+        const { token, id } = data;
 
-      return req.data;
+        const req = await axios.get(`/api/cart/items/${id}`, {
+          headers: { authorization: token },
+        });
+
+        return req.data;
+      } else {
+        const items =
+          JSON.parse(window.localStorage.getItem("cart-items-no-log")) || [];
+
+        console.log("ITEMS", items);
+
+        return items;
+      }
     } catch (err) {
       throw err;
     }
@@ -20,20 +32,34 @@ export const getItems = createAsyncThunk(
 
 export const addItem = createAsyncThunk("ADD_ITEM", async (data, thunkAPI) => {
   try {
-    const { token, cartInUse } = thunkAPI.getState();
+    const { token, cartInUse, isLogIn } = thunkAPI.getState();
 
-    const body = {
-      carritoId: cartInUse.id,
-      productId: data.id,
-    };
+    if (isLogIn) {
+      const body = {
+        carritoId: cartInUse.id,
+        productId: data.id,
+      };
 
-    const req = await axios.post("/api/cart/add", body, {
-      headers: { authorization: token },
-    });
+      const req = await axios.post("/api/cart/add", body, {
+        headers: { authorization: token },
+      });
 
-    const { productId, quantity } = req.data;
+      const { productId, quantity } = req.data;
+      return { productId, quantity };
+    } else {
+      const items =
+        JSON.parse(window.localStorage.getItem("cart-items-no-log")) || [];
 
-    return { productId, quantity };
+      const obj = { productId: data.id, quantity: 1 };
+
+      const newItems = [...items, obj];
+
+      window.localStorage.setItem(
+        "cart-items-no-log",
+        JSON.stringify(newItems)
+      );
+      return obj;
+    }
   } catch (err) {
     throw err;
   }
@@ -43,11 +69,23 @@ export const removeItem = createAsyncThunk(
   "REMOVE_ITEM",
   async (data, thunkAPI) => {
     try {
-      const { token, cartInUse } = thunkAPI.getState();
+      const { token, cartInUse, isLogIn } = thunkAPI.getState();
 
-      await axios.delete(`/api/cart/${cartInUse.id}/${data.id}`, {
-        headers: { authorization: token },
-      });
+      if (isLogIn) {
+        await axios.delete(`/api/cart/${cartInUse.id}/${data.id}`, {
+          headers: { authorization: token },
+        });
+      } else {
+        const items =
+          JSON.parse(window.localStorage.getItem("cart-items-no-log")) || [];
+
+        const newItems = items.filter((item) => item.productId !== data.id);
+
+        window.localStorage.setItem(
+          "cart-items-no-log",
+          JSON.stringify(newItems)
+        );
+      }
       return data;
     } catch (err) {
       throw err;
@@ -59,23 +97,37 @@ export const quantityRemove = createAsyncThunk(
   "QUANTITY_REMOVE_ITEM",
   async (data, thunkAPI) => {
     try {
-      const { token, cartInUse } = thunkAPI.getState();
+      const { token, cartInUse, isLogIn } = thunkAPI.getState();
+      if (isLogIn) {
+        const body = {
+          carritoId: cartInUse.id,
+          productId: data.id,
+          type: "remove",
+        };
 
-      const body = {
-        carritoId: cartInUse.id,
-        productId: data.id,
-        type: "remove",
-      };
+        await axios.put(`/api/cart`, body, {
+          headers: { authorization: token },
+        });
 
-      await axios.put(`/api/cart`, body, {
-        headers: { authorization: token },
-      });
+        const req = await axios.get(`/api/cart/items/${cartInUse.id}`, {
+          headers: { authorization: token },
+        });
 
-      const req = await axios.get(`/api/cart/items/${cartInUse.id}`, {
-        headers: { authorization: token },
-      });
+        return req.data;
+      } else {
+        const items = JSON.parse(
+          window.localStorage.getItem("cart-items-no-log")
+        );
 
-      return req.data;
+        const newItems = removeOneFromQuantity(items, data.id);
+
+        window.localStorage.setItem(
+          "cart-items-no-log",
+          JSON.stringify(newItems)
+        );
+        console.log("NEWITEMS", newItems);
+        return newItems;
+      }
     } catch (err) {
       throw err;
     }
@@ -86,23 +138,37 @@ export const quantityAdd = createAsyncThunk(
   "QUANTITY_ADD_ITEM",
   async (data, thunkAPI) => {
     try {
-      const { token, cartInUse } = thunkAPI.getState();
+      const { token, cartInUse, isLogIn } = thunkAPI.getState();
+      if (isLogIn) {
+        const body = {
+          carritoId: cartInUse.id,
+          productId: data.id,
+          type: "add",
+        };
 
-      const body = {
-        carritoId: cartInUse.id,
-        productId: data.id,
-        type: "add",
-      };
+        await axios.put(`/api/cart`, body, {
+          headers: { authorization: token },
+        });
 
-      await axios.put(`/api/cart`, body, {
-        headers: { authorization: token },
-      });
+        const req = await axios.get(`/api/cart/items/${cartInUse.id}`, {
+          headers: { authorization: token },
+        });
 
-      const req = await axios.get(`/api/cart/items/${cartInUse.id}`, {
-        headers: { authorization: token },
-      });
+        return req.data;
+      } else {
+        const items = JSON.parse(
+          window.localStorage.getItem("cart-items-no-log")
+        );
 
-      return req.data;
+        const newItems = addOneToQuantity(items, data.id);
+
+        window.localStorage.setItem(
+          "cart-items-no-log",
+          JSON.stringify(newItems)
+        );
+        console.log("NEWITEMS", newItems);
+        return newItems;
+      }
     } catch (err) {
       throw err;
     }
